pyproject.toml
[tool.poetry]
name = "nonebot-plugin-resolver"
version = "1.2.22"
description = "NoneBot2链接分享解析器插件。解析视频、图片链接/小程序插件，tiktok、bilibili、twitter等实时发送！"
authors = ["zhiyu1998 <renzhiyu0416@qq.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_resolver"}]
repository = "https://github.com/zhiyu1998/nonebot-plugin-resolver"
keywords = ["nonebot", "nonebot2", "resolver"]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.7"
httpx = ">=0.23"
PyExecJS = "^1.5.1"
bilibili-api-python = ">=16.2.0"
aiofiles = ">=0.8.0"
nonebot2 = "^2.4.0"
nonebot-adapter-onebot = "^2.4.6"
nonebot-plugin-htmlrender = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"